// src/type/types.ts
export enum Path {
  Home = 'home',
  Phones = 'phones',
  Tablets = 'tablets',
  Accessories = 'accessories',
  Favourites = 'favourites',
  Cart = 'cart',
}

export enum Sort {
  Newest = 'Newest',
  Alphabetically = 'Alphabetically',
  Cheapest = 'Cheapest',
}

export enum PerPage {
  Four = '4',
  Eight = '8',
  Sixteen = '16',
  All = 'All',
}